fix(cli): refactor file directories to make help command run
