--- stacks/cli/src/commands/bucket/index.ts
@@ -1,1 +1,0 @@
-export * from "./schema";
