--- stacks/cli/src/commands/bucket/interface.ts
@@ -1,47 +1,0 @@
-export interface BucketSchema {
-  _id: string;
-  title: string;
-  properties: Properties;
-  required?: string[];
-  [key: string]: any;
-}
-
-export type Properties = {[key: string]: Property};
-
-export type Property =
-  | BasicProperty
-  | ArrayProperty
-  | ObjectProperty
-  | RelationProperty
-  | LocationProperty;
-
-interface IProperty {
-  type: string;
-  enum?: any[];
-  [key: string]: any;
-}
-
-interface BasicProperty extends IProperty {
-  type: "string" | "textarea" | "color" | "richtext" | "storage" | "number" | "date" | "boolean";
-}
-
-interface ArrayProperty extends IProperty {
-  type: "array" | "multiselect";
-  items: Property;
-}
-
-interface ObjectProperty extends IProperty {
-  type: "object";
-  properties: Properties;
-  required?: string[];
-}
-
-interface RelationProperty extends IProperty {
-  type: "relation";
-  bucketId: string;
-  relationType: "onetoone" | "onetomany";
-}
-
-interface LocationProperty extends IProperty {
-  type: "location";
-}
