+++ stacks/cli/test/commands/bucket/schema.spec.ts
@@ -1,4 +1,4 @@
-import {createFileContent} from "@spica/cli";
+import {Schema} from "@spica/cli/src/commands/bucket/orm";
 
 function uglify(str) {
   return str.replace(/[\n ]+/g, "");
@@ -114,7 +114,7 @@ describe("ORM", () => {
   } as any;
 
   it("should create file content for bucket which includes all available types", () => {
-    const content = createFileContent([bucketAllTypes], "APIKEY", "APIURL", []);
+    const content = Schema.createFileContent([bucketAllTypes], "APIKEY", "APIURL", []);
     const expectation = `import * as Bucket from '@spica-devkit/bucket';
 /**
  * Call this method before interacting with buckets.
@@ -215,7 +215,12 @@ export namespace new_bucket {
     const bucket2 = {_id: "id2", title: "Users", properties: {name: {type: "string"}}};
 
     const warnings = [];
-    const content = createFileContent([bucket1, bucket2] as any, "APIKEY", "APIURL", warnings);
+    const content = Schema.createFileContent(
+      [bucket1, bucket2] as any,
+      "APIKEY",
+      "APIURL",
+      warnings
+    );
     const expectation = `import * as Bucket from '@spica-devkit/bucket';
 /**
  * Call this method before interacting with buckets.
